Move `rle` and `bitfield` to primitives; Re-org actor part
