+++ actor/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "plum_actor"
+version = "0.1.0"
+authors = ["The PolkaX Authors"]
+edition = "2018"
+license = "GPL-3.0"
+
+[dependencies]
+cid = { version = "0.5", features = ["impl-serde"] }
+minicbor = { version = "0.3", features = ["std"] }
+multihash = "0.11"
+serde = { version = "1.0", features = ["derive"] }
+serde_bytes = "0.11"
+serde_repr = "0.1"
+thiserror = "1.0"
+
+# ipfs
+#ipld-cbor = { package = "rust-ipld-cbor", git = "https://github.com/PolkaX/rust-ipfs", branch = "filecoin-master" }
+
+plum_address = { path = "../primitives/address" }
+plum_bigint = { path = "../primitives/bigint" }
+plum_bitfield = { path = "../primitives/bitfield" }
+plum_crypto = { path = "../crypto" }
+plum_hash = { path = "../primitives/hash" }
+plum_piece = { path = "../primitives/piece" }
+plum_sector = { path = "../primitives/sector" }
+plum_types = { path = "../primitives/types" }
+
+[dev-dependencies]
+hex-literal = "0.2"
