+++ actor/src/builtin/market/error.rs
@@ -1,6 +1,6 @@
-use thiserror::Error;
+// Copyright 2019-2020 PolkaX Authors. Licensed under GPL-3.0.
 
-#[derive(Debug, Error)]
+#[derive(Debug, thiserror::Error)]
 pub enum StorageMarketError {
     #[error("this proposal already do sign before")]
     AlreadySigned,
@@ -8,8 +8,8 @@ pub enum StorageMarketError {
     NotSigned,
     #[error("sign error: {0}")]
     Sign(#[from] Box<dyn std::error::Error + Sync + Send + 'static>),
-    #[error("sign error: {0}")]
-    Ipld(#[from] ipld_cbor::IpldCborError),
+    // #[error("sign error: {0}")]
+    // Ipld(#[from] ipld_cbor::IpldCborError),
     #[error("address error: {0}")]
     Address(#[from] plum_address::AddressError),
     #[error("crypto error: {0}")]
