+++ actor/src/builtin/market/state.rs
@@ -1,0 +1,1 @@
+// Copyright 2019-2020 PolkaX Authors. Licensed under GPL-3.0.
