+++ actor/src/builtin/miner/actor.rs
@@ -1,5 +1,7 @@
-use crate::abi::bitfield::BitField;
-use crate::abi::sector::RegisteredProof;
+// Copyright 2019-2020 PolkaX Authors. Licensed under GPL-3.0.
+
+use plum_bitfield::BitField;
+use plum_sector::RegisteredProof;
 
 pub enum CronEventType {
     WindowedPoStExpiration,
