+++ actor/src/builtin/miner/mod.rs
@@ -1,0 +1,6 @@
+// Copyright 2019-2020 PolkaX Authors. Licensed under GPL-3.0.
+
+pub mod actor;
+pub mod state;
+#[cfg(test)]
+pub mod tests;
