+++ actor/src/lib.rs
@@ -1,0 +1,3 @@
+// Copyright 2019-2020 PolkaX Authors. Licensed under GPL-3.0.
+
+pub mod builtin;
