--- chain/actor/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "plum_actor"
-version = "0.1.0"
-authors = ["The PolkaX Authors"]
-edition = "2018"
-license = "GPL-3.0"
-
-[dependencies]
-multihash = "0.11"
-cid = { version = "0.5", features = ["impl-serde"] }
-enum-repr-derive = "0.1.3"
-serde = { version = "1.0", features = ["derive"] }
-serde_bytes = "0.11"
-serde_repr = "0.1"
-serde_tuple = "0.4"
-serde_cbor = "0.11"
-thiserror = "1.0"
-
-# ipfs
-ipld-cbor = { package = "rust-ipld-cbor", git = "https://github.com/PolkaX/rust-ipfs", branch = "filecoin-master" }
-
-codec-rle = { path = "../../codec/rle" }
-plum_address = { path = "../../primitives/address" }
-plum_bigint = { path = "../../primitives/bigint" }
-plum_hash = { path = "../../primitives/hash" }
-plum_types = { path = "../../primitives/types" }
-plum_crypto = { path = "../../crypto" }
-
-[dev-dependencies]
-hex-literal = "0.2"
-serde_cbor = "0.11"
