--- chain/actor/src/abi/mod.rs
@@ -1,4 +1,0 @@
-pub mod bitfield;
-mod error;
-pub mod piece;
-pub mod sector;
