--- chain/actor/src/builtin/market/mod.rs
@@ -1,3 +1,0 @@
-pub mod actor;
-pub mod error;
-pub mod state;
