--- chain/actor/src/builtin/miner/mod.rs
@@ -1,4 +1,0 @@
-pub mod actor;
-pub mod state;
-#[cfg(test)]
-pub mod tests;
