--- chain/actor/src/builtin/mod.rs
@@ -1,2 +1,0 @@
-pub mod market;
-pub mod miner;
