--- chain/actor/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod abi;
-pub mod builtin;
