+++ primitives/bitfield/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "plum_bitfield"
+version = "0.1.0"
+authors = ["The PolkaX Authors"]
+edition = "2018"
+license = "GPL-3.0"
+
+[dependencies]
+minicbor = { version = "0.3", features = ["std"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_bytes = "0.11"
+
+rle = { path = "./rle" }
+
+plum_bytes = { path = "../bytes" }
+
+[dev-dependencies]
+serde_json = "1.0"
