+++ primitives/bitfield/rle/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "codec-rle"
+name = "rle"
 version = "0.1.0"
 authors = ["The PolkaX Authors"]
 edition = "2018"
